+++ app/models/tog/deprecation.rb
@@ -1,0 +1,8 @@
+module Deprecation # :nodoc:
+  extend ActiveSupport::Deprecation
+
+  def self.warn(message, callstack = caller)
+    message = '[TOG] ' + message.strip.gsub(/\s+/, ' ')
+    ActiveSupport::Deprecation.warn(message, callstack)
+  end
+end